requirements.txt
Flask
Flask-Cors
google-cloud-logging
google-cloud-bigquery
google-auth
python-dotenv